fix: compatibility with webpack v5
